requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
blinker==1.9.0
click==8.1.7
Django==5.1.6
djangorestframework==3.15.2
fastapi==0.115.8
Flask==3.1.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.2.0
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
Werkzeug==3.1.3
wheel==0.44.0